# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.6.0
  description: |
    Seshat API — Plataforma de gestão educacional com registro fotográfico, equipes, turmas e ocorrências.

    ## Banco de dados

    O PostgreSQL 16 roda **dentro do container de desenvolvimento** e é gerenciado
    automaticamente pelo `entrypoint.sh`:

    1. **Inicialização do cluster** — `pg_createcluster` é executado somente se o
       volume estiver vazio (`PG_VERSION` ausente em `PGDATA`).
    2. **Persistência** — dados armazenados no volume Docker `carometro-pg-dev`
       montado em `/var/lib/postgresql/16/main`.
    3. **Schema** — `drizzle-kit push` roda automaticamente **somente quando** o
       hash dos arquivos `lib/db/src/**/*.ts` muda. O hash é persistido em
       `$PGDATA/.carometro_schema_hash`.
    4. **Seed do administrador** — executado **apenas na primeira inicialização**
       via marcador `$PGDATA/.carometro_initialized`. O script `seed-admin.ts` é
       idempotente (não duplica o usuário se já existir).

    ### Comandos úteis

    ```
    docker compose run --rm dev db:push       # forçar reaplicação do schema
    docker compose run --rm dev db:seed       # verificar/recriar administrador
    docker compose run --rm dev db:hash-reset # remover marcadores e forçar re-setup
    docker compose down -v && docker compose up --build  # reinício limpo
    ```
servers:
  - url: /api
...
paths:
...
  /cursos:
    get:
      operationId: listCursos
      tags: [cursos]
      summary: List all cursos
      responses:
        "200":
          description: List of cursos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Curso"
    post:
      operationId: createCurso
      tags: [cursos]
      summary: Create a curso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CursoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Curso"
...
  /turmas:
    get:
      operationId: listTurmas
      tags: [turmas]
      summary: List all turmas
      parameters:
        - name: turnoId
          in: query
          required: false
          schema:
            type: string
        - name: cursoId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of turmas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Turma"
    post:
      operationId: createTurma
      tags: [turmas]
      summary: Create a turma
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TurmaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Turma"
...
  /estudantes:
    get:
      operationId: listEstudantes
      tags: [estudantes]
      summary: List all estudantes
      parameters:
        - name: turmaId
          in: query
          required: false
          schema:
            type: string
        - name: turnoId
          in: query
          required: false
          schema:
            type: string
        - name: cursoId
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of estudantes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Estudante"
    post:
      operationId: createEstudante
      tags: [estudantes]
      summary: Create an estudante
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstudanteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estudante"
...
  /import/cursos:
    post:
      operationId: importCursos
      tags: [import]
      summary: Import cursos from spreadsheet rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportCursosBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /import/turmas:
    post:
      operationId: importTurmas
      tags: [import]
      summary: Import turmas from spreadsheet rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportTurmasBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /import/estudantes:
    post:
      operationId: importEstudantes
      tags: [import]
      summary: Import estudantes from spreadsheet rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportEstudantesBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /import/disciplinas:
    post:
      operationId: importDisciplinas
      tags: [import]
      summary: Import disciplinas and their ofertas from spreadsheet rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportDisciplinasBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /import/professores:
    post:
      operationId: importProfessores
      tags: [import]
      summary: Import professores and their disciplina links from spreadsheet rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportProfessoresBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
  /disciplinas:
    get:
      operationId: listDisciplinas
      tags: [disciplinas]
      summary: List all disciplinas
      responses:
        "200":
          description: List of disciplinas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Disciplina"
    post:
      operationId: createDisciplina
      tags: [disciplinas]
      summary: Create a disciplina
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisciplinaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disciplina"
...
components:
  schemas:
...
    Curso:
      type: object
      required: [id, nome, criadoEm]
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        criadoEm:
          type: string
    CursoInput:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
          minLength: 1
    Turma:
      type: object
      required: [id, sigla, descricao, turnoId, turnoNome, cursoId, cursoNome, criadoEm]
      properties:
        id:
          type: string
          format: uuid
        sigla:
          type: string
        descricao:
          type: string
        turnoId:
          type: string
          format: uuid
        turnoNome:
          type: string
        cursoId:
          type: string
          format: uuid
        cursoNome:
          type: string
        criadoEm:
          type: string
    TurmaInput:
      type: object
      required: [sigla, descricao, turnoId, cursoId]
      properties:
        sigla:
          type: string
        descricao:
          type: string
        turnoId:
          type: string
          format: uuid
        cursoId:
          type: string
          format: uuid
    EstudanteEmail:
      type: object
      required: [id, email, tipo]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        tipo:
          type: string
          enum: [proprio, responsavel]
    EstudanteEmailInput:
      type: object
      required: [email, tipo]
      properties:
        email:
          type: string
        tipo:
          type: string
          enum: [proprio, responsavel]
    Estudante:
      type: object
      required: [id, nome, registro, turmaId, turmaSigla, turmaDescricao, turnoNome, cursoNome, criadoEm, emails]
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        registro:
          type: string
        fotoUrl:
          type: string
          nullable: true
        observacao:
          type: string
          nullable: true
        emails:
          type: array
          items:
            $ref: "#/components/schemas/EstudanteEmail"
        turmaId:
          type: string
          format: uuid
        turmaSigla:
          type: string
        turmaDescricao:
          type: string
        turnoNome:
          type: string
        cursoNome:
          type: string
        criadoEm:
          type: string
    EstudanteInput:
      type: object
      required: [nome, registro, turmaId]
      properties:
        nome:
          type: string
          minLength: 1
        registro:
          type: string
          minLength: 1
        observacao:
          type: string
          nullable: true
          maxLength: 300
        emails:
          type: array
          items:
            $ref: "#/components/schemas/EstudanteEmailInput"
        turmaId:
          type: string
          format: uuid
...
    ImportRow:
      type: object
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: true
    ImportCursosBody:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportRow"
    ImportTurmasBody:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportRow"
    ImportEstudantesBody:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportRow"
    ImportDisciplinasBody:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportRow"
    ImportProfessoresBody:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportRow"
    ImportResult:
      type: object
      required: [imported, errors]
      properties:
        imported:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    Disciplina:
      type: object
      required: [id, nome, criadoEm, ofertas]
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        criadoEm:
          type: string
        ofertas:
          type: array
          items:
            $ref: "#/components/schemas/DisciplinaOferta"
    DisciplinaOferta:
      type: object
      required: [id, disciplinaId, cursoId, cursoNome, turnoId, turnoNome]
      properties:
        id:
          type: string
          format: uuid
        disciplinaId:
          type: string
          format: uuid
        cursoId:
          type: string
          format: uuid
        cursoNome:
          type: string
        turnoId:
          type: string
          format: uuid
        turnoNome:
          type: string
    DisciplinaInput:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
          minLength: 1
...